backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0
google-cloud-aiplatform==1.38.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
python-dotenv==1.0.0==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.8.0
google-cloud-aiplatform>=1.38.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
